Add initial book addition form + styling
Right now the form div is just hovering there above everything else, need to display change corresponding to the buttons. also add functionality of submitting to js and not default form submission. finally need to make it so when clicking edit on a made book, the form is opened up with the details inside.